requirements.txt
fastapi>=0.110.0
pydantic==2.6.4
typing-extensions==4.9.0
groq>=0.5.0
python-dotenv>=1.0.0
httpx==0.27.2
uvicorn>=0.28.0
